+++ server/controllers/snippet.js
@@ -16,6 +16,10 @@ module.exports.deleteSnippet = async (accountId, snippetId, organizationId = nul
         return { code: 404, message: "Snippet does not exist" };
     }
 
+    if (snippet.sourceId) {
+        return { code: 403, message: "Cannot delete source-synced snippets" };
+    }
+
     await Snippet.destroy({ where: { id: snippetId } });
 };
 
@@ -28,6 +32,10 @@ module.exports.editSnippet = async (accountId, snippetId, configuration, organiz
 
     if (snippet === null) {
         return { code: 404, message: "Snippet does not exist" };
+    }
+
+    if (snippet.sourceId) {
+        return { code: 403, message: "Cannot edit source-synced snippets" };
     }
 
     const { organizationId: _, accountId: __, ...updateData } = configuration;
@@ -52,15 +60,27 @@ module.exports.listSnippets = async (accountId, organizationId = null) => {
     if (organizationId) {
         return await Snippet.findAll({ where: { organizationId } });
     }
-    return await Snippet.findAll({ where: { accountId, organizationId: null } });
+    return await Snippet.findAll({ where: { accountId, organizationId: null, sourceId: null } });
 };
 
 module.exports.listAllAccessibleSnippets = async (accountId, organizationIds = []) => {
     const whereClause = {
         [Op.or]: [
-            { accountId, organizationId: null },
+            { accountId, organizationId: null, sourceId: null },
             ...(organizationIds.length > 0 ? [{ organizationId: { [Op.in]: organizationIds } }] : [])
         ]
     };
     return await Snippet.findAll({ where: whereClause });
+};
+
+module.exports.listSourceSnippets = async (sourceId) => {
+    return await Snippet.findAll({ where: { sourceId } });
+};
+
+module.exports.listAllSourceSnippets = async () => {
+    return await Snippet.findAll({ 
+        where: { 
+            sourceId: { [Op.ne]: null } 
+        } 
+    });
 };
